% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.survForest.r
\name{predict.survForest}
\alias{predict.survForest}
\title{predict.survForest}
\usage{
\method{predict}{survForest}(object, testx, use.cores = 0, ...)
}
\arguments{
\item{object}{A fitted RLT object}

\item{testx}{Testing data}

\item{use.cores}{number of cores (by default use number of threads -1)}

\item{...}{...}
}
\value{
The predicted values.
}
\description{
The prediction function for personalized survival forest
}
\examples{
x = matrix(rnorm(1000), ncol = 10)
y = exp(rowMeans(x))
c = rbinom(100, 1, 0.9)
fit = survForest(x, y, c, ntrees = 10)
predict(fit, x)
}
