% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rfp_bcd_get_data.R
\name{rfp_bcd_get_data}
\alias{rfp_bcd_get_data}
\title{Convenience wrapper on bcdata::bcdc_query_geodata() and bcdata::collect() to retrieve Web Feature Service data from the BC Data Catalogue.}
\usage{
rfp_bcd_get_data(
  bcdata_record_id,
  col_extract = NULL,
  drop_geom = FALSE,
  col_filter = NULL,
  col_filter_value = NULL
)
}
\arguments{
\item{bcdata_record_id}{A character string specifying the BC Data Catalogue record ID permanent id (d0e5c8bc-7874-4c48-9c3e-431f772a250e),
name of the record (pscis-assessments) or \verb{object name} (ex. WHSE_FISH.PSCIS_ASSESSMENT_SVW)
can be used and can be found at https://catalogue.data.gov.bc.ca/}

\item{col_extract}{A character vector specifying which columns to select from the data. Cannot be NULL.}

\item{drop_geom}{A logical value indicating whether to drop the geometry column from the \code{sf} object. Default is FALSE}

\item{col_filter}{A character string specifying the column identifier to filter data on.}

\item{col_filter_value}{A character or character vector specifying the values to filter \code{col_filter} by.
If NULL (default), no filtering on \code{col_filter} is applied.}
}
\value{
A data frame with cleaned names and optionally an \code{sf} object.
}
\description{
This function queries geospatial data from the BC Data Catalogue using a specified record ID. It conditionally
filters the data based on a provided column identifier and value(s),
selects specific columns, cleans the column names. Geometry of object returned can be dropped if desired
}
\examples{
\dontrun{
rfp_bcd_get_data(bcdata_record_id = "d0e5c8bc-7874-4c48-9c3e-431f772a250e",
col_filter = 'FUNDING_PROJECT_NUMBER',
col_filter_value = 'skeena_2023_Phase1',
col_extract = c('EXTERNAL_CROSSING_REFERENCE', 'STREAM_CROSSING_ID'),
drop_geom = TRUE)
rfp_bcd_get_data(bcdata_record_id = "bc-airports", col_extract = c('AIRPORT_NAME', 'DESCRIPTION')) # Retrieves all records without filtering on col_filter
}
}
\seealso{
\code{\link{rfp_meta_bcd_xref}} for a function that extracts metadata from the BC Data Catalogue.
}
\concept{source bcdata}
