% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rfp_fs_backup.R
\name{rfp_fs_backup}
\alias{rfp_fs_backup}
\title{Backup a file with optional timestamp}
\usage{
rfp_fs_backup(
  path_dir_in,
  path_in_file,
  path_dir_out,
  stamp = format(Sys.time(), "\%Y\%m\%d\%H\%M")
)
}
\arguments{
\item{path_dir_in}{Character. Directory of the input file.}

\item{path_in_file}{Character. Name of the input file.}

\item{path_dir_out}{Character. Directory where the backup should be stored.}

\item{stamp}{Character or NULL. Optional string to append to the filename. Defaults to current timestamp.}
}
\value{
Character. Full path to the newly created backup file.
}
\description{
Copies a specified file to a backup directory, optionally appending a timestamp.
}
\examples{
\dontrun{
path_dir_in <- "~/Projects/gis/sern_peace_fwcp_2023"
path_in_file <- "background_layers.gpkg"
path_dir_out <- "~/Projects/gis/backups"
rfp_fs_backup(path_dir_in, path_in_file, path_dir_out)
}
}
