% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rfp_gpkg_to_geojson.R
\name{rfp_gpkg_to_geojson}
\alias{rfp_gpkg_to_geojson}
\title{Convert all layers in a GeoPackage to GeoJSON files}
\usage{
rfp_gpkg_to_geojson(
  dir_in = "~/test",
  dir_out = "~/test",
  file_name_in = "gps_2024"
)
}
\arguments{
\item{dir_in}{A character string representing the directory where the GeoPackage is located. Default is "~/test".}

\item{dir_out}{A character string representing the directory where the GeoJSON files will be saved. Default is "~/test".}

\item{file_name_in}{A character string for the name of the GeoPackage file (without extension). Default is "gps_2024".}
}
\description{
This function reads layers from a GeoPackage file and writes them as GeoJSON files.
We use geojsons to make the mapping convenient as they update automagically in QGIS without a restart and b/c
when in wsg84 they display by default on github and other web platforms
}
