% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rfp_meta_bcd_xref.R
\name{rfp_meta_bcd_xref}
\alias{rfp_meta_bcd_xref}
\title{Function to create a lookup table for BC Data Catalogue}
\usage{
rfp_meta_bcd_xref(pkg_exclude = c("bc-web-map-library"))
}
\arguments{
\item{pkg_exclude}{A character vector specifying the packages to exclude. Default is c("bc-web-map-library").
This package name is associated with several datasets that do not have wms download capabilities.}
}
\value{
A data frame with columns object_name, package_name, url_browser, description, id_package, id_dataset, url_download.
}
\description{
Function to create a lookup table for BC Data Catalogue
}
\examples{
\dontrun{
  lookup_table <- rfp_meta_bcd_xref()
  head(lookup_table)
}
}
\seealso{
Other meta bcdata: 
\code{\link{rfp_meta_bcd_xref_col_comments}()},
\code{\link{rfp_meta_bcd_xref_pkg_schtab}()}
}
\concept{meta bcdata}
