% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rfp_meta_bcd_xref_col_comments.R
\name{rfp_meta_bcd_xref_col_comments}
\alias{rfp_meta_bcd_xref_col_comments}
\title{rfp_meta_bcd_xref_col_comments}
\usage{
rfp_meta_bcd_xref_col_comments(schema_table, column)
}
\arguments{
\item{schema_table}{A string. The name of the schema_table to check.}

\item{column}{A string. The name of the column to check.}
}
\value{
A tibble with information about the column if it exists. Information included is the character value of
schema_table provided, the column name, the R column type, and the column comments.
}
\description{
Check if a given schema_table and column exist in the BC Data Catalogue WMS service.
If they exist, it returns a dataframe with information about the column.
}
\examples{
\dontrun{
rfp_meta_bcd_xref_col_comments("whse_imagery_and_base_maps.gsr_airports_svw", "airport_name")
}
}
\seealso{
Other meta bcdata: 
\code{\link{rfp_meta_bcd_xref}()},
\code{\link{rfp_meta_bcd_xref_pkg_schtab}()}
}
\concept{meta bcdata}
