% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rfp_source_bcdata.R
\name{rfp_source_bcdata}
\alias{rfp_source_bcdata}
\title{Download and Save BC Data to GPKG}
\usage{
rfp_source_bcdata(
  bcdata_record_id = NULL,
  path_gpkg = NULL,
  mask = NULL,
  layer_name = NULL
)
}
\arguments{
\item{bcdata_record_id}{\link{character} A character string specifying the BC Data Catalogue record ID permanent id (7ecfafa6-5e18-48cd-8d9b-eae5b5ea2881),
name of the record (pscis-assessments) or \verb{object name} (ex. WHSE_FISH.PSCIS_ASSESSMENT_SVW).  The name of the record
can be found at https://catalogue.data.gov.bc.ca/.  This does not need to be in capitals as we will convert to capitals
within the function.}

\item{path_gpkg}{\link{character} Path to the output GeoPackage file.}

\item{mask}{\link[sf:sf]{sf::sf} Optional masking polygon to clip spatial extent. Can massively increase download speed as first queries only bbox from api.}

\item{layer_name}{\link{character} Optional layer name for writing to the GPKG. Defaults to \code{bcdata_record_id}.}
}
\description{
Downloads a layer from the BC Data Catalogue, optionally intersects with a mask, and writes it to a GeoPackage.
}
