% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rfp_source_csv.R
\name{rfp_source_csv}
\alias{rfp_source_csv}
\title{Download and Convert CSVs to GeoPackage Layers}
\usage{
rfp_source_csv(path_gpkg, urls_csv)
}
\arguments{
\item{path_gpkg}{A string specifying the path to the output GeoPackage file.}

\item{urls_csv}{A character vector of URLs pointing to the CSV files to be downloaded and converted.}
}
\value{
Invisible NULL. This function is called for its side effects.
}
\description{
This function downloads CSV files from specified URLs and converts them into layers
within a GeoPackage using the \code{ogr2ogr} command-line tool from GDAL.
}
\examples{
\dontrun{
path_gpkg <- "background_layers.gpkg"
urls_csv <- c(
  "https://raw.githubusercontent.com/smnorris/bcfishpass/main/parameters/example_newgraph/parameters_habitat_method.csv",
  "https://raw.githubusercontent.com/smnorris/bcfishpass/main/parameters/example_newgraph/parameters_habitat_thresholds.csv"
)
rfp_source_csv(path_gpkg, urls_csv)
}
}
\seealso{
Other source: 
\code{\link{rfp_source_url}()}
}
\concept{source}
