% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rfp_source_url.R
\name{rfp_source_url}
\alias{rfp_source_url}
\title{Download and Convert Files to GeoPackage Layers with Optional Filtering}
\usage{
rfp_source_url(path_gpkg, urls, query = NULL)
}
\arguments{
\item{path_gpkg}{A string specifying the path to the output GeoPackage file.}

\item{urls}{A character vector of URLs pointing to the files to be downloaded and converted.}

\item{query}{An optional SQL query string to filter the data during the conversion process.
If NULL (default), no filter is applied.}
}
\value{
Invisible NULL. This function is called for its side effects.
}
\description{
This function downloads files from specified URLs and converts them into layers
within a GeoPackage using the \code{ogr2ogr} command-line tool from GDAL. It optionally
applies an SQL filter to the data during conversion if a query is specified.
}
\examples{
\dontrun{
path_gpkg <- "background_layers.gpkg"
urls <- c(
  "https://newgraph.s3.us-west-2.amazonaws.com/bcfishpass.crossings_vw.fgb",
  "https://raw.githubusercontent.com/smnorris/bcfishpass/main/parameters/example_newgraph/parameters_habitat_thresholds.csv"
)
query <- "watershed_group_code in ('ADMS', 'KLUM')"
rfp_source_url(path_gpkg, urls, query)
}
}
\seealso{
Other source: 
\code{\link{rfp_source_csv}()}
}
\concept{source}
