% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rfp_xref_fwa_wsg.R
\docType{data}
\name{rfp_xref_fwa_wsg}
\alias{rfp_xref_fwa_wsg}
\title{Table used to crosswalk freshwater atlas watershed group codes and watershed group names}
\format{
\subsection{\code{rfp_xref_fwa_wsg}}{

A data frame with 246 rows and 2 columns:
\describe{
\item{watershed_group_code}{The four character watershed group code, e.g. ADMS (Adams River), ALBN (Alberni Inlet).}
...
}
}
}
\source{
\url{https://catalogue.data.gov.bc.ca/dataset/freshwater-atlas-watershed-groups/resource/7239c84e-418a-4a9e-97bf-61b166410384}
}
\usage{
rfp_xref_fwa_wsg
}
\description{
Table that can be used to reliably get names of watershed groups from codes or the other way around.
Updateing procedures are documented in the \code{data-raw} directory.
}
\keyword{datasets}
